+++ src/Service/Constants/Command/GetConstantsCommand.php
@@ -4,17 +4,24 @@ namespace FluxIliasRestApi\Service\Constants\Command;
 
 use FluxIliasRestApi\Adapter\Constants\ConstantsDto;
 use ilObjOrgUnit;
+use ilObjUser;
 
 class GetConstantsCommand
 {
 
-    private function __construct() {
+    private function __construct(
+        private readonly ilObjUser $ilias_user
+    ) {
 
     }
 
 
-    public static function new() : static {
-        return new static();
+    public static function new(
+        ilObjUser $ilias_user
+    ) : static {
+        return new static(
+            $ilias_user
+        );
     }
 
 
@@ -29,6 +36,7 @@ class GetConstantsCommand
             USER_FOLDER_ID,
             SYSTEM_USER_ID,
             ANONYMOUS_USER_ID,
+            $this->ilias_user->getId(),
             ROLE_FOLDER_ID,
             SYSTEM_ROLE_ID,
             ANONYMOUS_ROLE_ID,