+++ src/Service/User/Command/UpdateAvatarCommand.php
@@ -57,7 +57,9 @@ class UpdateAvatarCommand
 
     private function updateAvatar(?UserDto $user, ?string $file) : ?UserIdDto
     {
-        if ($user === null || $user->id === $this->constants_service->getConstants()->root_user_id) {
+        $constants = $this->constants_service->getConstants();
+
+        if ($user === null || $user->id === $constants->root_user_id || $user->id === $constants->rest_user_user_id) {
             return null;
         }